Resolve "Landing Page"

See merge request c1623793/ons_group2!8

+++ src/main/java/com/ons/group2/ons_client_project/repository/UserSkillsRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface UserSkillsRepository extends JpaRepository<UserSkill, Integer> {
     Optional<UserSkill> findById(Integer id);
 
-    List<UserSkill> findAllByUser_Id(Integer userId);
+    List<UserSkill> findAllByUser_Id(Long userId);
 
-    void deleteByIdAndUser_Id(Integer skillId, Integer userId);
+    void deleteByIdAndUser_Id(Integer skillId, Long userId);
 }
